+++ src/pydash/exceptions.py
@@ -5,7 +5,9 @@ Exception classes.
 """
 
 
-class InvalidMethod(Exception):
+# NOTE: This needs to subclass AttributeError due to compatibility with typing.Protocol and
+#  runtime_checkable. See https://github.com/dgilland/pydash/issues/165
+class InvalidMethod(AttributeError):
     """
     Raised when an invalid pydash method is invoked through :func:`pydash.chaining.chain`.