+++ CHANGELOG.rst
@@ -4,6 +4,9 @@ Changelog
 =========
 
 
+- Fix compatibility issue between ``pydash.py_`` / ``pydash._`` and ``typing.Protocol`` + ``typing.runtime_checkable`` that caused an exception to be raised for ``isinstance(py_, SomeRuntimeCheckableProtocol)``.
+
+
 v5.0.1 (2021-06-27)
 -------------------